Rename utils.ts to config.ts
